+++ packages/contracts/src/support/htmlable.ts
@@ -1,0 +1,10 @@
+'use strict'
+
+export interface Htmlable {
+  /**
+   * Returns the content as an HTML string.
+   *
+   * @return string
+   */
+  toHtml(): string
+}