+++ packages/vite/src/contracts/plugin.ts
@@ -1,0 +1,36 @@
+'use strict'
+
+export interface PluginConfigContract {
+  /**
+   * The path or paths to the entrypoints to compile with Vite.
+   */
+  input: string | string[]
+
+  /**
+   * The "public" directory name.
+   *
+   * @default "public"
+   */
+  publicDirectory?: string
+
+  /**
+   * The "build" subdirectory name where compiled assets should be written.
+   *
+   * @default "build"
+   */
+  buildDirectory?: string
+
+  /**
+   * The SSR entry point path(s).
+   */
+  ssr?: string | string[]
+
+  /**
+   * The directory where the SSR bundle should be written.
+   *
+   * @default 'bootstrap/ssr'
+   */
+  ssrOutputDirectory?: string
+}
+
+export type DevServerUrl = `${'http' | 'https'}://${string}:${number}`